moved file actions to files module, added a bash script test, removed commented functions
